\name{Agro}
\alias{Agro}
\docType{data}
\title{
Agro: Mexican Household Income and Expenditure Survey 2010
}
\description{
Income and expense of all business related activities: agriculture, forestry, lumber, farming, fishing and hunting.
}
\usage{data("Agro")}
\format{
  \describe{
    \item{\code{Agro}}{Income and expense of all business related activities: agriculture, forestry, lumber, farming, fishing and hunting.}
}}
\source{ National Income and Expenditure Household Survey, ENIGH 2010 (new construction). Includes the 15 tables which are available at the INEGI portal and one additional table with AGEB data.
\url{www.inegi.org.mx}
}
\examples{
data(Agro)
}
\keyword{datasets}
