\name{Concen}
\alias{Concen}
\docType{data}
\title{
Concen: Mexican Household Income and Expenditure Survey 2010
}
\description{
Main household variables. The CONCENTRADO table includes all key variables to link all 14 tables.
}
\usage{data("enigh2010")}
\format{
  \describe{
    \item{\code{Concentrado}}{Main household variables. The CONCENTRADO table includes all key variables to link all 14 tables.}
  }
}
\source{ National Income and Expenditure Household Survey, ENIGH 2010 (new construction). Includes the 15 tables which are available at the INEGI portal and one additional table with AGEB data.
\url{www.inegi.org.mx}
}
\examples{
data("Concen")
}
\keyword{datasets}
