\name{Diseno}
\alias{Diseno}
\docType{data}
\title{
Diseno: Mexican Household Income and Expenditure Survey 2010
}
\description{
Diseno
}
\usage{data("Diseno")}
\format{
Household tables: A data frame with 130 observations on the following 9 variables.
  \describe{
    \item{\code{Diseno}}{...}
}}
\source{ National Income and Expenditure Household Survey, ENIGH 2010 (new construction). Includes the 15 tables which are available at the INEGI portal and one additional table with AGEB data.
\url{www.inegi.org.mx}
}
\examples{
data("Diseno")
}
\keyword{datasets}
