\name{G_costo}
\alias{G_costo}
\docType{data}
\title{
G_costo: Mexican Household Income and Expenditure Survey 2010
}
\description{
Expense in items bought with a loan. 
}
\usage{data("G_costo")}
\format{
  \describe{
    \item{\code{G_costo}}{...}
}}
\source{ National Income and Expenditure Household Survey, ENIGH 2010 (new construction). Includes the 15 tables which are available at the INEGI portal and one additional table with AGEB data.
\url{www.inegi.org.mx}
}
\examples{
data("G_costo")
}
\keyword{datasets}
