\name{G_diario}
\alias{G_diario}
\docType{data}
\title{
G_diario: Mexican Household Income and Expenditure Survey 2010
}
\description{
Household daily expense for food, beberages, tobacco and transit.
}
\usage{data("G_diario")}
\format{
Household tables: A data frame with 130 observations on the following 9 variables.
  \describe{
    \item{\code{Gastodiario}}{...}
}}
\source{ National Income and Expenditure Household Survey, ENIGH 2010 (new construction). Includes the 15 tables which are available at the INEGI portal and one additional table with AGEB data.
\url{www.inegi.org.mx}
}
\examples{
data("G_diario")
}
\keyword{datasets}
