\name{G_educa}
\alias{G_educa}
\docType{data}
\title{
G_educa: Mexican Household Income and Expenditure Survey 2010
}
\description{
Household expense in education related services and products.
}
\usage{data("G_educa")}
\format{
  \describe{
    \item{\code{Gastoeduca}}{...}
}}
\source{ National Income and Expenditure Household Survey, ENIGH 2010 (new construction). Includes the 15 tables which are available at the INEGI portal and one additional table with AGEB data.
\url{www.inegi.org.mx}
}
\examples{
data("G_educa")
}
\keyword{datasets}
