\name{G_tarje}
\alias{G_tarje}
\docType{data}
\title{
G_tarje: Mexican Household Income and Expenditure Survey 2010
}
\description{
Cost of the articles that were aquired by credit card.
}
\usage{data("enigh2010")}
\format{
  \describe{
    \item{\code{G_tarje}}{...}
}}
\source{ National Income and Expenditure Household Survey, ENIGH 2010 (new construction). Includes the 15 tables which are available at the INEGI portal and one additional table with AGEB data.
\url{www.inegi.org.mx}
}
\examples{
data("G_tarje")
}
\keyword{datasets}
