\name{Gastos}
\alias{Gastos}
\docType{data}
\title{
Gastos: Mexican Household Income and Expenditure Survey 2010
}
\description{
Household expenditure.
}
\usage{data("enigh2010")}
\format{
  \describe{
    \item{\code{Gastos}}{...}
}}
\source{ National Income and Expenditure Household Survey, ENIGH 2010 (new construction). Includes the 15 tables which are available at the INEGI portal and one additional table with AGEB data.
\url{www.inegi.org.mx}
}
\examples{
data("Gastos")
}
\keyword{datasets}
