\name{Hogares}
\alias{Hogares}
\docType{data}
\title{
Hogares: Mexican Household Income and Expenditure Survey 2010
}
\description{
Dwelling and household characteristics and expansion factor.
}
\usage{data("Hogares")}
\format{
  \describe{
    \item{\code{Hogares}}{...}
}}
\source{ National Income and Expenditure Household Survey, ENIGH 2010 (new construction). Includes the 15 tables which are available at the INEGI portal and one additional table with AGEB data.
\url{www.inegi.org.mx}
}
\examples{
data("Hogares")
}
\keyword{datasets}
