\name{Ingresos}
\alias{Ingresos}
\docType{data}
\title{
Ingresos: Mexican Household Income and Expenditure Survey 2010
}
\description{
Income and capital perceptions for all household members.
}
\usage{data("enigh2010")}
\format{
Household tables: A data frame with 130 observations on the following 9 variables.
  \describe{
    \item{\code{Ingresos}}{...}
}}
\source{ National Income and Expenditure Household Survey, ENIGH 2010 (new construction). Includes the 15 tables which are available at the INEGI portal and one additional table with AGEB data.
\url{www.inegi.org.mx}
}
\examples{
data("Ingresos")
}
\keyword{datasets}
