\name{Nomon}
\alias{Nomon}
\docType{data}
\title{Nomon: Mexican Household Income and Expenditure Survey 2010}
\description{Household and household members non-monetary income (or expenditure).}
\usage{data("Nomon")}
\format{
  \describe{
    \item{\code{Nomonetario}}{...}
}}
\source{ National Income and Expenditure Household Survey, ENIGH 2010 (new construction). Includes the 15 tables which are available at the INEGI portal and one additional table with AGEB data.
\url{www.inegi.org.mx}
}
\examples{
data("Nomon")
}
\keyword{datasets}
