\name{Pobla10}
\alias{Pobla10}
\docType{data}
\title{
Pobla10: Mexican Household Income and Expenditure Survey 2010
}
\description{
Demographic and social data for all household members.
}
\usage{data("Pobla10")}
\format{
  \describe{
    \item{\code{Pobla10}}{...}
}}
\source{ National Income and Expenditure Household Survey, ENIGH 2010 (new construction). Includes the 15 tables which are available at the INEGI portal and one additional table with AGEB data.
\url{www.inegi.org.mx}
}
\examples{
data("Pobla10")
}
\keyword{datasets}
