\name{Trabajos}
\alias{Trabajos}
\docType{data}
\title{
Trabajos: Mexican Household Income and Expenditure Survey 2010
}
\description{
Activities of all older than 12 household members.
}
\usage{data("Trabajos")}
\format{
  \describe{
    \item{\code{Trabajos}}{...}
}}
\source{ National Income and Expenditure Household Survey, ENIGH 2010 (new construction). Includes the 15 tables which are available at the INEGI portal and one additional table with AGEB data.
\url{www.inegi.org.mx}
}
\examples{
data("Trabajos")
}
\keyword{datasets}
